requirements.txt

fastapi
uvicorn
pandas
openpyxl
jinja2
requests
python-dotenv
reportlab
